\name{sortLines}
\alias{sortLines}
\title{Sort Selected Lines In Active Editor}
\usage{
sortLines()
}
\description{
Sorts selected lines in RStudio's active editor window. To be used as an add-in.
}
\author{
Dominic Comtois <dominic.comtois@gmail.com>
}
\note{
Credits to Gregory R. Warnes for the sort algorithm from \pkg{gtools} (\url{https://cran.r-project.org/web/packages/gtools/index.html}.
}



