\name{sortLinesD}
\alias{sortLinesD}
\title{Sort Selected Lines in Active Editor (Descending)}
\usage{
sortLinesD()
}
\description{
Sorts selected lines in RStudio's active editor window in descending order, using a number-smart algorithm. To be used as an add-in.
}
\author{
Dominic Comtois <dominic.comtois@gmail.com>
}
\note{
Credits to Gregory R. Warnes for the sort algorithm from \pkg{gtools} (\url{https://cran.r-project.org/web/packages/gtools/index.html}.
}



